refactor(cli): extract check runners and output pipeline
Convert cli.py to cli/ package. Extract output pipeline (_emit_findings,
_format_coverage_line, _resolve_format, _output_and_exit) to _output.py
and git helpers + check runners (_get_git_diff, _get_git_blame,
_write_timing, _run_*) to _runners.py.

Submodules access check functions, reporting, and stdlib modules through
the parent package (import docvet.cli as _cli_pkg) so that test mocks
on docvet.cli.* take effect — 100+ mocker.patch calls work unchanged.

- Reduce cli/__init__.py from 1,456 to 994 lines (32% reduction)
- New _output.py at 215 lines, _runners.py at 336 lines
- Entry point docvet.cli:app resolves unchanged

Closes #373
